requirements.txt
flask
pymongo
certifi
pypdf
transformers
torch
sentence-transformers
python-dotenv
accelerate
tiktoken
protobuf
waitress
seaborn
icrawler
reportlab
qrcode
Flask-Mail
gunicorn
